+++ Leaf/Http/Request.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace Leaf\Http;
 
 /**
@@ -195,6 +196,18 @@ class Request
         return $union;
     }
 
+    protected function sanitize($data)
+    {
+        if (is_array($data)) {
+            foreach ($data as $key => $value) {
+                $data[$this->sanitize($key)] = $this->sanitize($value);
+            }
+        } else {
+            $data = htmlspecialchars($data, ENT_QUOTES, 'UTF-8');
+        }
+        return $data;
+    }
+
     /**
      * Returns request data
      *
@@ -204,46 +217,17 @@ class Request
      *
      * @param  string           $key
      */
-    public function get($param) {
-        if ($this->requestMethod == "POST" || $this->requestMethod == "PUT" || $this->requestMethod == "PATCH" || $this->requestMethod == "DELETE") {
-            if (isset($_POST[$param])) {
-                return htmlspecialchars($_POST[$param], ENT_QUOTES, 'UTF-8');
-            } else {
-                $data = json_decode($this->request, true);
-                return isset($data[$param]) ? htmlspecialchars($data[$param], ENT_QUOTES, 'UTF-8') : null;
-            }
-        } else {
-            return isset($_GET[$param]) ? htmlspecialchars($_GET[$param], ENT_QUOTES, 'UTF-8') : null;
-        }
+    public function get($param)
+    {
+        return $this->sanitize($this->body()[$param]);
     }
 
     /**
      * Get all the response data as an associative array
      */
-    public function body() {
-        $data = json_decode($this->request, true);
-
-        $body = array();
-
-        if($this->requestMethod === "GET") {
-            foreach($_GET as $key => $value) {
-                $body[$key] = htmlspecialchars($value, ENT_QUOTES, 'UTF-8');
-            }
-            return count($body) > 0 ? $body : null;
-        }
-        if ($this->requestMethod == "POST" || $this->requestMethod == "PUT" || $this->requestMethod == "PATCH" || $this->requestMethod == "DELETE") {
-            if (isset($_POST) && !empty($_POST)) {
-                foreach($_POST as $key => $value) {
-                    $body[$key] = htmlspecialchars($value, ENT_QUOTES, 'UTF-8');
-                }
-                return count($body) > 0 ? $body : null;
-            } else {
-                foreach($data as $key => $value) {
-                    $body[$key] = htmlspecialchars($value, ENT_QUOTES, 'UTF-8');
-                }
-                return count($body) > 0 ? $body : null;
-            }
-        }
+    public function body()
+    {
+        return $this->sanitize($_GET + $_POST + json_decode($this->request, true));
     }
 
     /**
@@ -401,7 +385,7 @@ class Request
     public function getHost()
     {
         if (isset($this->env['HTTP_HOST'])) {
-            if(preg_match('/^(\[[a-fA-F0-9:.]+\])(:\d+)?\z/', $this->env['HTTP_HOST'], $matches)) {
+            if (preg_match('/^(\[[a-fA-F0-9:.]+\])(:\d+)?\z/', $this->env['HTTP_HOST'], $matches)) {
                 return $matches[1];
             } else {
                 if (strpos($this->env['HTTP_HOST'], ':') !== false) {
@@ -432,7 +416,7 @@ class Request
      */
     public function getPort()
     {
-        return (int)$this->env['SERVER_PORT'];
+        return (int) $this->env['SERVER_PORT'];
     }
 
     /**