+++ Leaf/Wynter/Exceptions/ComponentNotFoundException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Leaf\Wynter\Exceptions;
-
-class ComponentNotFoundException extends \Exception
-{
-    use BypassViewHandler;
-}