+++ Leaf/Wynter/Macros/wynter-view.blade.php
@@ -1,5 +1,0 @@
-@extends($layout)
-
-@section($section)
-    @wynter($component, $componentParameters)
-@endsection