+++ CHANGELOG.md
@@ -1,14 +1,37 @@
 <!-- markdownlint-disable no-duplicate-header -->
 # Changelog
 
+## v2.1.x - 26th June, 2020
+
+### Changed
+
+- Added fix for `Leaf\Http\Request` on "array" type objects
+
+## v2.1.0 - 22nd June, 2020
+
+### Added
+
+- Added `Leaf\Auth::auto_connect`
+- Added default bypass for CORS errors
+- Added `Mysqli::auto_connect`
+- Added optional `db_type` option to `Leaf\Db\PDO` connection
+- Added `PDO::auto_connect`
+- Added new db module (`Leaf\Db`)
+
+### Changed
+
+- Switched default db to `Leaf\Db`
+
+### Removed
+
+- Removed Leaf\Wynter
+
 ## v2.1.0 - alpha - 24th May, 2020
 
 ### Added
 
 - Added Route::resource
 - Added Session::retrieve
-
-### Fixed
 
 ### Changed